common/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
alembic>=1.12.0
pydantic[email]>=2.0.0
aio-pika>=9.3.0
redis>=5.0.1
prometheus-client>=0.17.1
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation-sqlalchemy>=0.41b0
opentelemetry-instrumentation-httpx>=0.41b0
opentelemetry-instrumentation-aio-pika>=0.41b0
opentelemetry-instrumentation-redis>=0.41b0
opentelemetry-instrumentation-aiohttp-client>=0.41b0
# Service discovery dependencies
python-consul>=1.1.0 
aiohttp>=3.8.5
tenacity>=8.2.3
httpx>=0.24.1
setuptools>=65.6.3
